+++ react-memo/src/components/ReactGuideCopyright.js
@@ -1,0 +1,17 @@
+import { Component } from 'react';
+
+
+export default class ReactGuideCopyright extends Component {
+  render() {
+    const date = new Date();
+
+    return (
+      <>
+        <hr />
+        <p className='text-center'>
+          Copyright &copy; {date.getFullYear()} <em>ReactMemo</em> all right reserved.
+        </p>
+      </>
+    );
+  }
+};