+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
 
+## v22.02.1 (2022-02-13)
+- create-react-app을 사용한 스터디 샘플 (#3)
+- useState, useEffect 사용해보기 (#4)
+- react를 사용한 샘플 CRUD 기능 구현 (#5)
+
 ## v22.02.0 (2022-02-02)
 - CDN을 사용한 리액트 기본 학습